+++ CHANGELOG.md
@@ -29,6 +29,7 @@ following conventions:
 ### Breaking
 
 * Add `maybe` functions
+* Add `prefixError` functions
 
 ## 1.1.1.1 (2022-03-01)